handle feedback from review of @dwhswenson
